LayoutTests/fast/mediastream/RTCPeerConnection-statsSelector-expected.txt

Tests the RTCPeerConnection stats selector argument.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS pc = new RTCPeerConnection({iceServers:[{urls:'stun:foo.com'}]}) did not throw exception.
PASS navigator.mediaDevices.getUserMedia({audio:true, video:true}).then(gotStream, error) did not throw exception.
PASS Got a stream.
PASS statsHandler2 was called
PASS result.size is >= 1
PASS timestamp is >= startTime
PASS local.kind is "video"
PASS successfullyParsed is true

TEST COMPLETE
